FIX: ex00 illegal comparison

+++ cpp_00/ex00/main.cpp
@@ -6,7 +6,7 @@
 /*   By: jwillert <jwillert@student.42heilbronn.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/01 15:43:17 by jwillert          #+#    #+#             */
-/*   Updated: 2023/08/04 14:35:10 by jwillert         ###   ########.fr       */
+/*   Updated: 2023/08/04 18:12:13 by jwillert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	main(int argc, char **argv)
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
 		return (0);
 	}
-	for(size_t i = 1; i < argc; ++i)
+	for(size_t i = 1; i < (size_t) argc; ++i)
 	{
 		std::string	str;
